\name{kml-internal}
% \alias{[,LongData-method}
% \alias{[,Partition-method}
% \alias{[<-,LongData-method}
% \alias{[<-,Partition-method}
% \alias{plot,LongData-method}
% \alias{plotTraj,LongData-method}
% \alias{plotSubGroups,LongData-method}
% \alias{as.clusterizLongData,LongData-method}
% \alias{selectSupTrajSizeMin}
% \alias{selectSupTrajSizeMin,LongData-method}
\alias{kmlFast}
\alias{plotLegend}
\alias{legendCol}
\alias{expandStartingCond}
\alias{cutScreen}
\alias{fastOrSlow}
\alias{exportPartition}
\alias{choiceChangeParam}
\alias{partPermut}
\alias{exportPartition}
\alias{exportPartition,ClusterLongData,numeric}
\alias{exportPartition,ClusterLongData,numeric-method}
\alias{ClusterLongData_choice}
\alias{[,ParChoice,ANY-method}
\alias{[<-,ParChoice,ANY,ANY,ANY-method}

\title{~ Internal KmL objects and methods ~}
\description{Internal: KmL objects and methods}
\details{These are not to be called by the user.\cr
  If you have a specific need or for more details on KmL, feel free to contact
  the author
  (who, even if he is very lazy , will be very happy to answer and - why not? -
  start a new collaboration...)
}





\seealso{
  Overview: \code{\link{kml-package}} \cr
}


\keyword{internal}

