\name{kml-internal}
\alias{[,LongData-method}
\alias{[,Partition-method}
\alias{[<-,LongData-method}
\alias{[<-,Partition-method}
\alias{plot,LongData-method}
\alias{plotTraj,LongData-method}
\alias{plotSubGroups,LongData-method}
\alias{as.clusterizLongData,LongData-method}
\alias{selectSupTrajSizeMin}
\alias{selectSupTrajSizeMin,LongData-method}
\alias{expandPartition}
\alias{expandPartition,Partition-method}
\alias{toto}

\title{~ Internal kml objects and methods ~}
\description{Internal: KmL objects and methods}
\details{These are not to be called by the user.\cr
  If you have a specific need or for more details on KmL, feel free to contact
  the author
  (which is a hard-worker, will be very happy to answer and - why not? -
  start a new collaboration...)
}


\references{Article submited\cr
  Web site: \code{\link[http://christophe.genolini.free.fr/kml]{http://christophe.genolini.free.fr/kml}}
}

\section{Author(s)}{
  Christophe Genolini\cr
  PSIGIAM: Paris Sud Innovation Group in Adolescent Mental Health\cr
  INSERM U669 / Maison de Solenn / Paris\cr\cr

  Contact author : <genolini@u-paris10.fr>
}

\section{English translation}{
  Raphaël Ricaud\cr
  Laboratoire "Sport & Culture" / "Sports & Culture" Laboratory \cr
  University of Paris 10 / Nanterre
}

\seealso{
  Overview: \code{\link{kml-package}} \cr
  Classes : \code{\linkS4class{ClusterizLongData}}, \code{\linkS4class{Clusterization}}, \code{\linkS4class{ArtificialLongData}} \cr
  Methods : \code{\link{clusterizLongData}}, \code{\link{kml}}, \code{\link{generateArtificialLongData}}, \code{\link{choice}}, \code{\link{as.clusterizLongData}}\cr
  Plot : \code{\link[kml:plot]{plot: overview}}, \code{\link[=plotTraj]{plot(ClusterizLongData)}},
  \code{\link[=plotCalinski]{plot(Calinski)}},
  \code{\link[=plotSubGroups]{plotSubGroups(ClusterizLongData)}}, \code{\link[=plotAll]{plotAll(ClusterizLongData)}}
}


\keyword{internal}

