\name{kml-internal}
\alias{addClusterization}
\alias{addPartition}
%\alias{as.cld}
%\alias{as.clusterizLongData}
\alias{as.longData}
\alias{as.longData.ArtificialLongData}
\alias{as.longData.data.frame}
\alias{as.longData.matrix}
%\alias{choice}
\alias{choicePlot}
\alias{cld}
\alias{cleanProg}
\alias{clusterization}
\alias{clusterizLongData}
\alias{clusters}
\alias{expandPartition}
\alias{exportClusterization}
\alias{getCalinskiCriterion}
\alias{getClusterizList}
\alias{getClusters}
\alias{getConvergenceTime}
\alias{getDetWithin}
\alias{getId}
\alias{getNbClusters}
\alias{getPercentEachCluster}
\alias{getTime}
\alias{getTraceBetween}
\alias{getTraceWithin}
\alias{getTraj}
\alias{getTrajSizeMin}
\alias{getVarBetween}
\alias{getVarName}
\alias{getVarWithin}
\alias{importClusterization}
\alias{imputeLongData}
%\alias{kml}
\alias{kml1}
\alias{ld}
\alias{longData}
\alias{nbClusters}
\alias{nbId}
\alias{nbTime}
\alias{partition}
\alias{plot.Criterion}
\alias{selectTrajNoNA}
\alias{setClusters<-}
\alias{setGenericVerif}
\alias{setId<-}
\alias{setNbClusters<-}
\alias{setTime<-}
\alias{setTraj<-}
\alias{setTrajSizeMin<-}
\alias{setVarName<-}
\alias{trajNoNA}
\alias{imputeLongData,matrix,ANY,factor-method}
\alias{kml1,matrix-method}
\alias{longData,missing,missing,missing,missing,missing-method}
\alias{longData,ANY,ANY,ANY,ANY,ANY-method}
\alias{longData,ANY,ANY,ANY,ANY,missing-method}
\alias{longData,ANY,ANY,ANY,missing,ANY-method}
\alias{longData,ANY,ANY,ANY,missing,missing-method}
\alias{partition,missing,missing,missing-method}
\alias{partition,ANY,ANY,ANY-method}



\title{Internal kml objects}
\description{~ Internal: KmL objects ~}
\details{These are not to be called by the user.\cr
For a specific need, for more details on KmL, feel free to contact
the author
(which is a hard-worker, will be very happy to answer and - why not? - start a new
collaboration...}
\keyword{internal}

