\name{kml-internal}
\alias{[,LongData-method}
\alias{[,Partition-method}
\alias{[<-,LongData-method}
\alias{[<-,Partition-method}
\alias{plot,LongData-method}
\alias{plotTraj,LongData-method}
\alias{plotSubGroups,LongData-method}
\alias{as.clusterizLongData,LongData-method}
\alias{selectSupTrajSizeMin}
\alias{selectSupTrajSizeMin,LongData-method}
\alias{expandPartition}
\alias{expandPartition,Partition-method}
\alias{toto}

\title{~ Internal kml objects and methods ~}
\description{Internal: KmL objects and methods}
\details{These are not to be called by the user.\cr
  If you have a specific need or for more details on KmL, feel free to contact
  the author
  (who, even if he is very lazy , will be very happy to answer and - why not? -
  start a new collaboration...)
}


\references{Article "KmL: K-means for Longitudinal Data", in
  Computational Statistics (accepted on 11-11-2009) \cr
  Web site: \url{http://christophe.genolini.free.fr/kml}
}
\section{Author(s)}{
  Christophe Genolini\cr
  PSIGIAM: Paris Sud Innovation Group in Adolescent Mental Health\cr
  INSERM U669 / Maison de Solenn / Paris\cr\cr

  Contact author : \email{genolini@u-paris10.fr}
}

\section{English translation}{
  Raphaël Ricaud\cr
  Laboratoire "Sport & Culture" / "Sports & Culture" Laboratory \cr
  University of Paris 10 / Nanterre
}





\seealso{
  Overview: \code{\link{kml-package}} \cr
  Classes : \code{\linkS4class{ClusterizLongData}}, \code{\linkS4class{Clusterization}} \cr
  Methods : \code{\link{clusterizLongData}}, \code{\link{kml}}, \code{\link{choice}}, \code{\link{as.clusterizLongData}}\cr
  Plot : \code{\link[kml:plot]{plot(ClusterizLongData)}},
  \code{\link[=plotCriterion]{plotCriterion}},
  \code{\link[=plotSubGroups]{plotSubGroups(ClusterizLongData)}}, \code{\link[=plotAll]{plotAll(ClusterizLongData)}}
}


\keyword{internal}

